requirements.txt
# Core Dependencies
aiohttp>=3.9.0
asyncpg>=0.29.0
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
websockets>=12.0
redis>=5.0.0

# Database
asyncpg>=0.29.0  # PostgreSQL/TimescaleDB
redis>=5.0.0  # Redis for caching

# Data Analysis
pandas>=2.1.0
numpy>=1.24.0

# WebSocket
websockets>=12.0
aiohttp>=3.9.0

# Configuration
python-dotenv>=1.0.0

# Optional: Binance API (if using python-binance)
# python-binance>=1.0.19

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Dashboard (Optional)
rich>=13.7.0
